backend/requirements.txt
Python3.7
web browser
FLask1.0.2